Factor our applytheme() function
